requirements.txt
langchain-text-splitters
google-genai==1.47.0
faiss-cpu
numpy==1.26.4
transformers==4.28.0
sentence-transformers==4.57.6
uvicorn==0.27.0
torch==2.1.0
fastapi==0.112.0
